Fixing database

+++ TrabalhoAPI/Controllers/CourseController.cs
@@ -2,6 +2,7 @@
 using Context.Enum;
 using Context.Models;
 using Context.Relations;
+using Context.Services;
 using Context.Tools;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
@@ -29,37 +30,28 @@ namespace TrabalhoAPI.Controllers
         {
 
             var result = new ResultData<Course>() { Error = true, Status = HttpStatusCode.BadRequest };
-            Course course;
             try
             {
-                if (Regex.IsMatch(course_name, @"[a-zA-Z\s]+"))
-                {
-                    course = new Course(course_name, status);
-                }
-                else
-                {
-                    throw new ArgumentException("Nome inválido");
-                }
-                using (db)
-                {
+                CourseServices.ValidateName(course_name);
+                Course course = new Course(course_name, status);
+                if (db.Curso.Where(q => q.Name.ToLower() == course_name.ToLower()).Any())
+                    throw new ArgumentException($"O nome {course.Name} já esta cadastrado");
 
-                    if (db.Curso.Where(q => q.Name.ToLower() == course_name.ToLower()).Any())
-                    {
-                        throw new ArgumentException($"O nome {course.Name} já esta cadastrado");
-                    }
-
-                    db.Curso.Add(course);
-                    db.SaveChanges();
-                    result.Error = false;
-                    result.Status = HttpStatusCode.OK;
-                    result.Data = db.Curso.ToList();
-                    return Ok(result);
-                }
+                db.Curso.Add(course);
+                db.SaveChanges();
+                result.Error = false;
+                result.Status = HttpStatusCode.OK;
+                result.Data = db.Curso.ToList();
+                return Ok(result);
             }
             catch (Exception e)
             {
                 result.Message.Add(e.Message);
                 return BadRequest(result);
+            }
+            finally
+            {
+                db.Dispose();
             }
         }
 
@@ -69,45 +61,48 @@ namespace TrabalhoAPI.Controllers
         public ActionResult AddSubject(int courseId, int subjectId)
         {
             var result = new ResultData<Course>() { Error = true, Status = HttpStatusCode.BadRequest };
-            using (db)
+
+            try
             {
-                try
+                var course = db.Curso.Where(q => q.IdCourse == courseId).FirstOrDefault();
+                var subject = db.Materia.Where(q => q.IdSubject == subjectId).FirstOrDefault();
+
+                if (course == null || subject == null)
                 {
-                    var course = db.Curso.Where(q => q.IdCourse == courseId).FirstOrDefault();
-                    var subject = db.Materia.Where(q => q.IdSubject == subjectId).FirstOrDefault();
-
-                    if (course == null || subject == null)
-                    {
-                        throw new ArgumentException("Item não encontrado");
-                    }
-                    if (subject.Status == SubjectStatus.Inactive)
-                    {
-                        throw new ArgumentException("Materia com Status 'INATIVO' não pode ser adicionada ao curso.");
-                    }
-
-                    course.Enrollments.Add(new CourseSubject()
-                    {
-                        Subject = subject,
-                        Course = course
-                    });
-
-                    db.SaveChanges();
-
-                    result.Error = false;
-                    result.Message.Add("Matéria adicionada com sucesso");
-                    result.Status = HttpStatusCode.OK;
-                    result.Data = db.Curso.Include(q => q.Enrollments).ToList();
-                    return Ok(result);
-
+                    throw new ArgumentException("Item não encontrado");
                 }
-                catch (Exception e)
+                if (subject.Status == SubjectStatus.Inactive)
                 {
-                    result.Error = true;
-                    result.Message.Add(e.Message);
-                    result.Status = HttpStatusCode.BadRequest;
-                    return BadRequest(result);
+                    throw new ArgumentException("Materia com Status 'INATIVO' não pode ser adicionada ao curso.");
                 }
+
+                course.Enrollments.Add(new CourseSubject()
+                {
+                    Subject = subject,
+                    Course = course
+                });
+
+                db.SaveChanges();
+
+                result.Error = false;
+                result.Message.Add("Matéria adicionada com sucesso");
+                result.Status = HttpStatusCode.OK;
+                result.Data = db.Curso.Include(q => q.Enrollments).ToList();
+                return Ok(result);
+
             }
+            catch (Exception e)
+            {
+                result.Error = true;
+                result.Message.Add(e.Message);
+                result.Status = HttpStatusCode.BadRequest;
+                return BadRequest(result);
+            }
+            finally
+            {
+                db.Dispose();
+            }
+
         }
 
         [HttpPost]
@@ -140,7 +135,7 @@ namespace TrabalhoAPI.Controllers
                     db.SaveChanges();
 
                     result.Error = false;
-                    result.Message.Add("Matéria adicionada com sucesso");
+                    result.Message.Add("Aluno adicionado com sucesso");
                     result.Status = HttpStatusCode.OK;
                     result.Data = db.Curso.Include(q => q.Enrollments).ToList();
                     return Ok(result);
@@ -216,21 +211,22 @@ namespace TrabalhoAPI.Controllers
                 var result = new ResultData<Course>();
                 try
                 {
-                    using (db)
+                    if (!db.Curso.Where(q => q.IdCourse == id).Any())
                     {
-                        if (!db.Curso.Select(q => q.Enrollments).Any() && !db.Curso.Select(q => q.CourseContains).Any())
-                        {
-                            var course = db.Curso.Where(q => q.IdCourse == id).FirstOrDefault();
-                            db.Curso.Remove(course);
-                            db.SaveChanges();
-                            result.Error = false;
-                            result.Data = db.Curso.ToList();
-                            result.Message.Add("OK");
-                            result.Status = HttpStatusCode.OK;
-                            return Ok(result);
-                        }
-                        throw new ArgumentException("O curso possui materias ou alunos");
+                        throw new ArgumentException("O curso não existe");
                     }
+                    if (!db.Curso.Select(q => q.Enrollments).Any() && !db.Curso.Select(q => q.CourseContains).Any())
+                    {
+                        var course = db.Curso.Where(q => q.IdCourse == id).FirstOrDefault();
+                        db.Curso.Remove(course);
+                        db.SaveChanges();
+                        result.Error = false;
+                        result.Data = db.Curso.ToList();
+                        result.Message.Add("OK");
+                        result.Status = HttpStatusCode.OK;
+                        return Ok(result);
+                    }
+                    throw new ArgumentException("O curso possui materias ou alunos");
                 }
                 catch (Exception e)
                 {
@@ -238,7 +234,89 @@ namespace TrabalhoAPI.Controllers
                     result.Message.Add(e.Message);
                     return BadRequest(result);
                 }
+                finally
+                {
+                    db.Dispose();
+                }
             }
+        }
+
+        [HttpDelete]
+        [Route("DeleteSubject")]
+        public ActionResult DeleteSubject(int id, int subjectId)
+        {
+            {
+                var result = new ResultData<Course>();
+                try
+                {
+                    var subject = db.Curso.SelectMany(q => q.Enrollments).Where(q => q.IdSubject == subjectId).FirstOrDefault();
+                    var hasStudents = db.Materia.SelectMany(q => q.Allocation).Where(q => q.IdSubject == subjectId).Any();
+                    if (subject is null)
+                    {
+                        throw new ArgumentException("A materia não existe");
+                    }
+                    if (!hasStudents)
+                    {
+                        var course = db.Curso.Where(q => q.IdCourse == id).Select(q => q.Enrollments).FirstOrDefault();
+                        course.Remove(subject);
+                        db.SaveChanges();
+                        result.Error = false;
+                        result.Data = db.Curso.ToList();
+                        result.Message.Add("OK");
+                        result.Status = HttpStatusCode.OK;
+                        return Ok(result);
+                    }
+                    throw new ArgumentException("A materia possui alunos");
+                }
+                catch (Exception e)
+                {
+                    result.Error = true;
+                    result.Message.Add(e.Message);
+                    return BadRequest(result);
+                }
+                finally
+                {
+                    db.Dispose();
+                }
+            }
+        }
+        [HttpDelete]
+        [Route("DeleteStudent")]
+        public ActionResult DeleteStudent(int id, string studentId)
+        {
+
+            var result = new ResultData<Course>();
+            try
+            {
+                var student = db.Curso.SelectMany(q => q.CourseContains).Where(q => q.User.Role == Occupation.Student && q.IdUser == studentId).FirstOrDefault();
+                if (student is null)
+                {
+                    throw new ArgumentException("O aluno não existe");
+                }
+                if (!db.Curso.Select(q => q.Enrollments).Any() && !db.Curso.Select(q => q.CourseContains).Any())
+                {
+                    var course = db.Curso.Where(q => q.IdCourse == id).Select(q => q.CourseContains).FirstOrDefault();
+                    course.Remove(student);
+                    db.SaveChanges();
+                    result.Error = false;
+                    result.Data = db.Curso.ToList();
+                    result.Message.Add("OK");
+                    result.Status = HttpStatusCode.OK;
+                    return Ok(result);
+                }
+                throw new ArgumentException("O curso possui materias ou alunos");
+            }
+            catch (Exception e)
+            {
+                result.Error = true;
+                result.Message.Add(e.Message);
+                return BadRequest(result);
+            }
+            finally
+            {
+                db.Dispose();
+            }
+
         }
     }
 }
